requirements.txt
pyyaml>=6.0
requests>=2.28.0
rich>=13.0.0
click>=8.1.0